+++ src/config/create-missions.ts
@@ -6,7 +6,9 @@ dotenv.config();
 
 async function createMissions() {
   try {
-    await mongoose.connect(process.env.MONGO_URI);
+    if (!process.env.MONGO_URI) throw new Error("Datenbank connection missing");
+    const MONGO_URI = process.env.MONGO_URI;
+    await mongoose.connect(MONGO_URI);
     await Mission.deleteMany({});
     console.log("✅ Alle Missionen gelöscht");
